modified

+++ src/qmi-nas.h
@@ -111,10 +111,10 @@ qmi_nas_get_sig_strength_output_unref(QmiNasGetSigStrengthOutput *output);
 
 gboolean
 qmi_nas_get_sig_strength_output_get_result(QmiNasGetSigStrengthOutput *output, GError **error);
-const guint16 qmi_nas_get_sig_strength_output_get_rsrp(QmiNasGetSigStrengthOutput *output);
-const guint16  qmi_nas_get_sig_strength_output_get_snr(QmiNasGetSigStrengthOutput *output);
-const guint8 qmi_nas_get_sig_strength_output_get_radio_if(QmiNasGetSigStrengthOutput *output);
-const guint8 qmi_nas_get_sig_strength_output_get_rsrq(QmiNasGetSigStrengthOutput *output);
+const gint16 qmi_nas_get_sig_strength_output_get_rsrp(QmiNasGetSigStrengthOutput *output);
+const float  qmi_nas_get_sig_strength_output_get_snr(QmiNasGetSigStrengthOutput *output);
+const gint8 qmi_nas_get_sig_strength_output_get_radio_if(QmiNasGetSigStrengthOutput *output);
+const gint8 qmi_nas_get_sig_strength_output_get_rsrq(QmiNasGetSigStrengthOutput *output);
